fixed betty errors on dog.h and changed int age to float age

+++ 0x0E-structures_typedef/dog.h
@@ -1,10 +1,15 @@
 #ifndef DOG_H
 #define DOG_H
-
+/**
+* struct dog - structure of dog name, age anf owner
+* @name: name element
+* @age: age element
+* @owner: owner element
+*/
 struct dog
 {
 	char *name;
-	int age;
+	float age;
 	char *owner;
 };
 #endif
